General tidy up and improvements
* Off by one error (thanks @HenningJ)
* Ignore clipboard contents from specific apps (and if the apps write a blank item to the pasteboard)
* Fix warnings
* 64 bit-ify

+++ QSPasteboardProxyView.m
@@ -91,7 +91,7 @@
 
 
 
-- (unsigned int)draggingSourceOperationMaskForLocal:(BOOL)isLocal{
+- (NSUInteger)draggingSourceOperationMaskForLocal:(BOOL)isLocal{
         return NSDragOperationEvery;
 }
